tests: Move recent RT tests to GPU-AV
Only left `BasicTraceRays` in core validation test, for the sake of
having a more easy to find ray tracing test template. The gpu-av tests
now run with core validation, so positive coverage should only be
improved
Also renamed `ray_tracing_gpu` and `ray_tracing_gpu_nv`
