+++ tests/CMakeLists.txt
@@ -102,6 +102,9 @@ target_sources(vk_layer_validation_tests PRIVATE
     unit/gpu_av_spirv.cpp
     unit/gpu_av_spirv_positive.cpp
     unit/gpu_av_ray_query_positive.cpp
+    unit/gpu_av_ray_tracing_nv.cpp
+    unit/gpu_av_ray_tracing.cpp
+    unit/gpu_av_ray_tracing_positive.cpp
     unit/graphics_library.cpp
     unit/graphics_library_positive.cpp
     unit/host_image_copy.cpp
@@ -143,8 +146,6 @@ target_sources(vk_layer_validation_tests PRIVATE
     unit/query_positive.cpp
     unit/ray_tracing.cpp
     unit/ray_tracing_nv.cpp
-    unit/ray_tracing_gpu_nv.cpp
-    unit/ray_tracing_gpu.cpp
     unit/ray_tracing_pipeline.cpp
     unit/ray_tracing_pipeline_positive.cpp
     unit/ray_tracing_pipeline_nv.cpp
